refactor: relocate ColumnSliceProfile to slicing module (DAT-183 step 1)
Move ColumnSliceProfile model from analysis/quality_summary/db_models.py
to analysis/slicing/db_models.py where it semantically belongs. This
prepares for deleting the quality_summary module in the next step.
