+++ src/dataraum/analysis/quality_summary/__init__.py
@@ -7,9 +7,9 @@ aggregating findings across all slices of that column.
 from dataraum.analysis.quality_summary.agent import QualitySummaryAgent
 from dataraum.analysis.quality_summary.db_models import (
     ColumnQualityReport,
-    ColumnSliceProfile,
 )
 from dataraum.analysis.quality_summary.processor import summarize_quality
+from dataraum.analysis.slicing.db_models import ColumnSliceProfile
 
 __all__ = [
     # Main entry point
