+++ src/dataraum/analysis/quality_summary/db_models.py
@@ -80,63 +80,6 @@ class ColumnQualityReport(Base):
     slice_column: Mapped[Column] = relationship(foreign_keys=[slice_column_id])
 
 
-class ColumnSliceProfile(Base):
-    """Stores quality metrics for a column within a specific slice.
-
-    Each record represents the quality profile of one column in one slice,
-    enabling analysis of quality patterns across slices.
-    """
-
-    __tablename__ = "column_slice_profiles"
-    __table_args__ = (
-        Index("idx_slice_profiles_source_column", "source_column_id"),
-        Index("idx_slice_profiles_slice_column", "slice_column_id"),
-        Index("idx_slice_profiles_lookup", "source_column_id", "slice_column_id", "slice_value"),
-    )
-
-    profile_id: Mapped[str] = mapped_column(String, primary_key=True, default=lambda: str(uuid4()))
-
-    # Reference to source column (in the original typed table)
-    source_column_id: Mapped[str] = mapped_column(
-        ForeignKey("columns.column_id", ondelete="CASCADE"), nullable=False
-    )
-
-    # Reference to slice definition column
-    slice_column_id: Mapped[str] = mapped_column(
-        ForeignKey("columns.column_id", ondelete="CASCADE"), nullable=False
-    )
-
-    # Identifiers
-    source_table_name: Mapped[str] = mapped_column(String, nullable=False)
-    column_name: Mapped[str] = mapped_column(String, nullable=False)
-    slice_column_name: Mapped[str] = mapped_column(String, nullable=False)
-    slice_value: Mapped[str] = mapped_column(String, nullable=False)
-
-    # Basic metrics
-    row_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    null_ratio: Mapped[float | None] = mapped_column(Float)
-    distinct_count: Mapped[int | None] = mapped_column(Integer)
-
-    # Quality indicators
-    quality_score: Mapped[float | None] = mapped_column(Float)
-    has_issues: Mapped[bool] = mapped_column(Integer, nullable=False, default=False)  # SQLite bool
-    issue_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
-    # Extended metrics stored as JSON for flexibility
-    # Contains: min_value, max_value, mean_value, stddev, benford_*, outlier_*, etc.
-    profile_data: Mapped[dict[str, Any] | None] = mapped_column(JSON)
-
-    # Timestamps
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False, default=lambda: datetime.now(UTC)
-    )
-
-    # Relationships
-    source_column: Mapped[Column] = relationship(foreign_keys=[source_column_id])
-    slice_column: Mapped[Column] = relationship(foreign_keys=[slice_column_id])
-
-
 __all__ = [
     "ColumnQualityReport",
-    "ColumnSliceProfile",
 ]
